+++ demo/Assets/SuperAwesome/Classes/SAInterstitialAd.cs
@@ -44,7 +44,7 @@ namespace SuperAwesome {
 		private static SAConfiguration configuration	= SuperAwesome.getInstance().defaultConfiguration ();
 
 		// instance constructor
-		private static void tryAndCreateOnce () {
+		public static void createInstance () {
 			// create just one static instance for ever!
 			if (staticInstance == null) {
 				GameObject obj = new GameObject ();
@@ -71,20 +71,6 @@ namespace SuperAwesome {
 			}
 		}
 
-		// MonoDevelop start implementation
-		void Start (){
-			if (this.GetComponent<Image> () != null) {
-				Color current = this.GetComponent<Image>().color;
-				current.a = 0;
-				this.GetComponent<Image>().color = current;
-			}
-		}
-		
-		// MonoDevelop update implementation
-		void Update () {
-			// do nothing
-		}
-
 		////////////////////////////////////////////////////////////////////
 		// Interstitial specific method
 		////////////////////////////////////////////////////////////////////
@@ -92,7 +78,7 @@ namespace SuperAwesome {
 		public static void load (int placementId) {
 
 			// create an instrance of an SAInterstitialAd (for callbacks)
-			tryAndCreateOnce ();
+			createInstance ();
 
 #if (UNITY_IPHONE && !UNITY_EDITOR) 
 			SAInterstitialAd.SuperAwesomeUnitySAInterstitialAdLoad(placementId, 
@@ -121,7 +107,7 @@ namespace SuperAwesome {
 		public static void play (int placementId) {
 
 			// create an instrance of an SAInterstitialAd (for callbacks)
-			tryAndCreateOnce ();
+			createInstance ();
 
 #if (UNITY_IPHONE && !UNITY_EDITOR) 
 			SAInterstitialAd.SuperAwesomeUnitySAInterstitialAdPlay(placementId,
@@ -151,7 +137,7 @@ namespace SuperAwesome {
 		public static bool hasAdAvailable (int placementId) {
 
 			// create an instrance of an SAInterstitialAd (for callbacks)
-			tryAndCreateOnce ();
+			createInstance ();
 
 #if (UNITY_IPHONE && !UNITY_EDITOR) 
 			return SAInterstitialAd.SuperAwesomeUnitySAInterstitialAdHasAdAvailable(placementId);