+++ demo/Assets/SuperAwesome/Classes/SAVideoAd.cs
@@ -50,7 +50,7 @@ namespace SuperAwesome {
 		private static SAConfiguration configuration 		= SuperAwesome.getInstance().defaultConfiguration ();
 
 		// instance constructor
-		private static void tryAndCreateOnce () {
+		private static void createInstance () {
 			// create just one static instance for ever!
 			if (staticInstance == null) {
 				GameObject obj = new GameObject ();
@@ -75,20 +75,6 @@ namespace SuperAwesome {
 #endif
 			}
 		}
-
-		// MonoDevelop start implementation
-		void Start (){
-			if (this.GetComponent<Image> () != null) {
-				Color current = this.GetComponent<Image>().color;
-				current.a = 0;
-				this.GetComponent<Image>().color = current;
-			}
-		}
-		
-		// MonoDevelop update implementation
-		void Update () {
-			// do nothing
-		}
 		
 		////////////////////////////////////////////////////////////////////
 		// Interstitial specific method
@@ -97,7 +83,7 @@ namespace SuperAwesome {
 		public static void load (int placementId) {
 
 			// create an instrance of an SAVideoAd (for callbacks)
-			tryAndCreateOnce ();
+			createInstance ();
 
 #if (UNITY_IPHONE && !UNITY_EDITOR) 
 			SAVideoAd.SuperAwesomeUnitySAVideoAdLoad(placementId, 
@@ -126,7 +112,7 @@ namespace SuperAwesome {
 		public static void play (int placementId) {
 
 			// create an instrance of an SAVideoAd (for callbacks)
-			tryAndCreateOnce ();
+			createInstance ();
 
 #if (UNITY_IPHONE && !UNITY_EDITOR) 
 			SAVideoAd.SuperAwesomeUnitySAVideoAdPlay(placementId,
@@ -162,7 +148,7 @@ namespace SuperAwesome {
 		public static bool hasAdAvailable (int placementId) {
 
 			// create an instrance of an SAVideoAd (for callbacks)
-			tryAndCreateOnce ();
+			createInstance ();
 
 #if (UNITY_IPHONE && !UNITY_EDITOR) 
 			return SAVideoAd.SuperAwesomeUnitySAVideoAdHasAdAvailable(placementId);